% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq.R
\name{square}
\alias{square}
\title{It use to square a value}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value is calculated}
}
\value{
The square of the input
}
\description{
It use to square a value
}
\examples{
square(10)
}
